Remove -Minline flag. (#573)
This was needed for old PGI versions (circa 16.x), but it does not seem
to be important anymore. The documentation also suggests that the syntax
of this option has changed and the old formulation is no longer valid.
`-Minline` also seems to have strange semantics when it is passed
alongside `-MD` and `-MT` -- as is done by modern CMake versions for
dependency tracking.